pyproject.toml
[tool.poetry]
name = "cinemaforum"
version = "0.1.0"
description = ""
authors = ["chugunova24 <t.chugunova24@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
celery = "^5.3.6"
sqlalchemy = "^2.0.25"
uvicorn = "^0.25.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
redis = "^5.0.1"
flower = "^2.0.1"
pytest = "^7.4.4"
httpx = "^0.26.0"
allure-pytest = "^2.13.2"
sqlalchemy-utils = "^0.41.1"
fastapi-restful = "^0.5.0"
typing-inspect = "^0.9.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
flake8 = "^7.0.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
websockets = "^12.0"
jinja2 = "^3.1.3"
asyncio = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
